+++ scripts/release/build-commands/run-automated-tests.js
@@ -2,23 +2,7 @@
 
 'use strict';
 
-const chalk = require('chalk');
-const {exec} = require('child-process-promise');
-const {logPromise} = require('../utils');
-
-const runYarnTask = async (cwd, task, errorMessage) => {
-  try {
-    await exec(`yarn ${task}`, {cwd});
-  } catch (error) {
-    throw Error(
-      chalk`
-      ${errorMessage}
-
-      {white ${error.stdout}}
-    `
-    );
-  }
-};
+const {logPromise, runYarnTask} = require('../utils');
 
 module.exports = async ({cwd}) => {
   await logPromise(runYarnTask(cwd, 'lint', 'Lint failed'), 'Running ESLint');
@@ -27,8 +11,13 @@ module.exports = async ({cwd}) => {
     'Running Flow checks'
   );
   await logPromise(
-    runYarnTask(cwd, 'test', 'Jest failed'),
-    'Running Jest tests',
+    runYarnTask(cwd, 'test', 'Jest tests failed in development'),
+    'Running Jest tests in the development environment',
+    true
+  );
+  await logPromise(
+    runYarnTask(cwd, 'test-prod', 'Jest tests failed in production'),
+    'Running Jest tests in the production environment',
     true
   );
 };
